+++ src/main/java/in/gagan/base/framework/controller/UserController.java
@@ -25,6 +25,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 
+import in.gagan.base.framework.controller.base.AbstractController;
 import in.gagan.base.framework.entity.user.User;
 import in.gagan.base.framework.util.DTOMapper;
 import org.springframework.beans.factory.annotation.Autowired;
