+++ src/main/java/in/gagan/base/framework/controller/base/AbstractController.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.controller;
+package in.gagan.base.framework.controller.base;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
